src/graphql updated for sqlite
