+++ ts/src/core/query/custom_query_sqlite.test.ts
@@ -1,4 +1,3 @@
-import { QueryRecorder } from "../../testutils/db_mock";
 import { Viewer } from "../base";
 import {
   FakeUser,
