+++ ts/src/graphql/query/assoc_edge_connection.test.ts
@@ -1,13 +1,9 @@
 import { Pool } from "pg";
-import { Data } from "../../core/base";
 import { QueryRecorder } from "../../testutils/db_mock";
-
-import {
-  FakeUser,
-  UserToContactsFkeyQuery,
-} from "../../testutils/fake_data/index";
+import { FakeUser, UserToContactsQuery } from "../../testutils/fake_data/index";
 import { createEdges } from "../../testutils/fake_data/test_helpers";
 import { commonTests } from "./shared_edge_connection";
+import { sharedAssocTests } from "./shared_assoc_test";
 jest.mock("pg");
 QueryRecorder.mockPool(Pool);
 
@@ -18,7 +14,9 @@ beforeEach(async () => {
 });
 
 commonTests({
-  getQuery: (v, user: FakeUser) => UserToContactsFkeyQuery.query(v, user),
-  tableName: "fake_contacts",
-  sortCol: "created_at",
+  getQuery: (v, user: FakeUser) => new UserToContactsQuery(v, user),
+  tableName: "user_to_contacts_table",
+  sortCol: "time",
 });
+
+sharedAssocTests();
