+++ ts/src/graphql/query/custom_edge_connection_sqlite.test.ts
@@ -1,22 +1,19 @@
-import { Pool } from "pg";
-import { Data } from "../../core/base";
-import { QueryRecorder } from "../../testutils/db_mock";
-
+import { setupSqlite } from "../../testutils/db/test_db";
 import {
   FakeUser,
   UserToContactsFkeyQuery,
 } from "../../testutils/fake_data/index";
-import { createEdges } from "../../testutils/fake_data/test_helpers";
+import {
+  createEdges,
+  tempDBTables,
+} from "../../testutils/fake_data/test_helpers";
 import { commonTests } from "./shared_edge_connection";
-jest.mock("pg");
-QueryRecorder.mockPool(Pool);
 
 beforeEach(async () => {
-  QueryRecorder.clear();
   await createEdges();
-  QueryRecorder.clearQueries();
 });
 
+setupSqlite(`sqlite:///custom_edge_connection.db`, tempDBTables);
 commonTests({
   getQuery: (v, user: FakeUser) => UserToContactsFkeyQuery.query(v, user),
   tableName: "fake_contacts",
